# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 74520518 进入技术交流群 请备注，谢谢！",
        "title": "kubetabs API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/wenjianzhang/kubetabs/blob/master/LICENSE.md"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/sd/cpu": {
            "get": {
                "description": "CPU 使用量 DiskCheck checks the disk usage.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "summary": "CPU 使用量",
                "responses": {
                    "200": {
                        "description": "OK - Load average: 2.39, 2.13, 1.97 | Cores: 2",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "WARNING",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "CRITICAL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sd/disk": {
            "get": {
                "description": "服务器硬盘使用量 DiskCheck checks the disk usage.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "summary": "服务器硬盘使用量",
                "responses": {
                    "200": {
                        "description": "OK - Free space: 16321MB (15GB) / 51200MB (50GB) | Used: 31%",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "WARNING",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "CRITICAL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/sd/os": {
            "get": {
                "description": "Os",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "summary": "OS",
                "responses": {
                    "200": {
                        "description": "OK - Free space: 455MB (0GB) / 8192MB (8GB) | Used: 5%",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "WARNING",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "CRITICAL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sd/ram": {
            "get": {
                "description": "内存使用量 RAMCheck checks the disk usage.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "summary": "内存使用量",
                "responses": {
                    "200": {
                        "description": "OK - Free space: 455MB (0GB) / 8192MB (8GB) | Used: 5%",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "WARNING",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "CRITICAL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}